requirements-test.txt
black==19.10b0
flake8==3.8.3
flake8-broken-line==0.2.0
flake8-bugbear==20.1.4
flake8-comprehensions==3.2.3
flake8-eradicate==0.4.0
flake8-print==3.1.4
gevent==20.6.2
isort==4.3.21
mypy==0.781
pdbpp==0.10.2
pydocstyle==5.0.2
pytest==5.4.3
pytest-cov==2.10.0
pytest-freezegun==0.4.1
pytest-mock==3.1.1
requests-mock==1.8.0
wf-bandit-tests==1.0.1
